Speed up Apecs (#100)
* The coroutine implementation of apecs was a big performance hit, so I have changed that to use classic iterators. This has given massive gains in performance but required a few changes to sprocket, in particular the lua glue code.
* Given that views over different types now have different types (previously all `apx::generator<spkt::entity>`, each view needed its own set of functions rather than just a `New`.
* However, the whole thing could be simplified to just needing two functions each, which could be templatised on the type, actually making the implementation much smaller and clearer.
* Now that the `view` functions in lua are only implemented using two C++ functions, it is less error prone and most likely a lot faster.
* Fix some warnings in `CircadianCycle` by fully operating with floats.
* Simplify the header file for the A* path finding, as well as making some of the implementation more readable (more to go).
* Next task will be to look into combining apecs `view_t` and `view_iterator`. The view could potentially store the state and `begin` could return itself, which may make things simpler.

+++ Game/CircadianCycle.cpp
@@ -48,31 +48,31 @@ void CircadianCycle::SetTime(int hours, int mins, int secs)
     Modulo();
 }
 
-void CircadianCycle::AddSeconds(double seconds)
+void CircadianCycle::AddSeconds(float seconds)
 {
     d_seconds += seconds;
     Modulo();
 }
 
-double CircadianCycle::GetSeconds() const
+float CircadianCycle::GetSeconds() const
 {
     return d_seconds; 
 }
 
-void CircadianCycle::SetSeconds(double seconds)
+void CircadianCycle::SetSeconds(float seconds)
 {
     d_seconds = seconds;
     Modulo();
 }
 
-double CircadianCycle::GetAngle() const
+float CircadianCycle::GetAngle() const
 { 
-    return glm::two_pi<double>() * d_seconds / SECONDS_IN_DAY;
+    return glm::two_pi<float>() * d_seconds / SECONDS_IN_DAY;
 }
 
-void CircadianCycle::SetAngle(double angle)
+void CircadianCycle::SetAngle(float angle)
 {
-    d_seconds = (angle / glm::two_pi<double>()) * SECONDS_IN_DAY;
+    d_seconds = (angle / glm::two_pi<float>()) * SECONDS_IN_DAY;
     Modulo();
 }
 
@@ -113,7 +113,7 @@ std::string CircadianCycle::ToString(bool twelveHour) const
 {
     int hours = 0;
     int mins = 0;
-    double seconds = d_seconds;
+    float seconds = d_seconds;
     while (seconds > 3600.0f) { ++hours; seconds -= 3600.0f; }
     while (seconds > 60.0f) { ++mins; seconds -= 60.0f; }
